requirements.txt
requests>=2.28.0
PyPDF2>=3.0.0
beautifulsoup4>=4.11.0
langdetect>=1.0.9
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Flask>=3.0.0
Flask-CORS>=4.0.0

# Image Search Dependencies
qdrant-client>=1.7.0
pymongo>=4.6.0
pdf2image>=1.16.3
Pillow>=10.0.0
numpy>=1.24.0
PyMuPDF>=1.23.0

# Embedding Providers
cohere>=5.0.0
openai>=1.0.0

# Authentication
bcrypt>=4.0.0
PyJWT>=2.8.0
python-docx>=1.2.0
